requirements.txt
streamlit==1.29.0
pypdf==3.17.1
pdfplumber==0.10.0
openai==1.6.1
langchain==0.1.0
plotly==5.17.0
pandas==2.1.4
python-dotenv==1.0.0 